+++ multimodal/src/autogluon/multimodal/utils/__init__.py
@@ -53,6 +53,7 @@ from .log import (
     get_fit_start_message,
     get_gpu_message,
     make_exp_dir,
+    handle_deprecated_args,
 )
 from .matcher import compute_semantic_similarity, convert_data_for_ranking, create_siamese_model, semantic_search
 from .metric import compute_ranking_score, compute_score, get_minmax_mode, get_stopping_threshold, infer_metrics, infer_problem_type_by_eval_metric
